requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
configparser>=6.0.0
aiohttp>=3.8.4
aiohttp_jinja2>=1.5.0
aiohttp_session>=2.12.0
cryptography>=40.0.2
pytz
psutil